Changes3

+++ src/main/java/com/driver/services/BlogService.java
@@ -26,28 +26,30 @@ public class BlogService {
 
     public List<Blog> showBlogs(){
         //find all blogs
-        List<Blog> blogList = new ArrayList<>();
-
-        return blogRepository1.saveAll(blogList);
+        return  blogRepository1.findAll();
 
     }
 
     public void createAndReturnBlog(Integer userId, String title, String content) {
-
-
         //create a blog at the current time
-        Blog blog = new Blog(title, content);
-        User user = userRepository1.findById(userId).get();
-        List<Blog> listofblogs = user.getBlogList();
-        listofblogs.add(blog);
-        user.setBlogList(listofblogs);
+        Blog blog=new Blog(title,content,new Date());
+
+        //Updating the userInformation and changing its blogs
+        blog.setUser(userRepository1.findById(userId).get());
+
+        User user=userRepository1.findById(userId).get();
+
+        List<Blog> res=user.getBlogList();
+
+        res.add((Blog) blogRepository1);
+
+        user.setBlogList(res);
 
         blogRepository1.save(blog);
         userRepository1.save(user);
-
         //updating the blog details
 
-        //Updating the userInformation and changing its blogs
+
 
     }
 
@@ -58,18 +60,28 @@ public class BlogService {
 
     public void addImage(Integer blogId, String description, String dimensions){
         //add an image to the blog after creating it
-
-        Image image=new Image(description,dimensions);
         Blog blog=blogRepository1.findById(blogId).get();
-        List<Image> AllImages=blog.getImageList();
-        AllImages.add(image);
-        blog.setImageList(AllImages);
 
+        Image image=imageService1.createAndReturn(blog,description,dimensions);
+        image.setBlog(blog);
+
+        List<Image> imageList=blog.getImageList();
+
+        if(imageList==null)
+            imageList=new ArrayList<>();
+
+        imageList.add(image);
+
+        blog.setImageList(imageList);
         blogRepository1.save(blog);
+
     }
 
     public void deleteBlog(int blogId){
         //delete blog and corresponding images
+
+        if(blogRepository1.findById(blogId).get()==null) return;
+
         blogRepository1.deleteById(blogId);
     }
 }
